+++ Example/VialerSIPLib/VialerSIPLib_Example-Bridging-Header.h
@@ -4,6 +4,9 @@
 
 #import "DDLogWrapper.h"
 #import <VialerSIPLib/VSLAudioController.h>
+#import <VialerSIPLib/VSLCall.h>
 #import <VialerSIPLib/VialerSIPLib.h>
 #import <VialerSIPLib/VSLRingtone.h>
 #import <VialerSIPLib/CallKitProviderDelegate.h>
+#import <VialerSIPLib/VSLAudioCodecs.h>
+#import <VialerSIPLib/VSLVideoCodecs.h>