+++ Pod/Classes/Configurations/VSLEndpointConfiguration.m
@@ -10,7 +10,7 @@
 
 static NSUInteger const VSLEndpointConfigurationMaxCalls = 4;
 static NSUInteger const VSLEndpointConfigurationLogLevel = 5;
-static NSUInteger const  VSLEndpointConfigurationLogConsoleLevel = 4;
+static NSUInteger const VSLEndpointConfigurationLogConsoleLevel = 4;
 static NSString * const VSLEndpointConfigurationLogFileName = nil;
 static NSUInteger const VSLEndpointConfigurationClockRate = PJSUA_DEFAULT_CLOCK_RATE;
 static NSUInteger const VSLEndpointConfigurationSndClockRate = 0;