+++ Pod/Classes/VSLEndpoint.h
@@ -54,7 +54,11 @@ typedef NS_ENUM(NSInteger, VSLEndpointState) {
     /**
      *  Endpoint is running.
      */
-    VSLEndpointStarted
+    VSLEndpointStarted,
+    /**
+     *  Endpoint is closing.
+     */
+    VSLEndpointClosing,
 };
 #define VSLEndpointStateString(VSLEndpointState) [@[@"VSLEndpointStopped", @"VSLEndpointStarting", @"VSLEndpointStarted"] objectAtIndex:VSLEndpointState]
 
@@ -158,6 +162,6 @@ typedef NS_ENUM(NSInteger, VSLEndpointState) {
  *
  *  @param activate BOOL, if YES, only iLBC will be used.
  */
-- (void)onlyUseILBC:(BOOL)activate;
+- (void)onlyUseILBC:(BOOL)activate __attribute__((deprecated("Deprecated, use VSLCodecConfigurarion to add codecs to the endpoint instead")));
 
 @end