+++ Pod/Classes/VSLEndpoint.m
@@ -11,10 +11,12 @@
 #import "VialerSIPLib.h"
 #import "VSLCall.h"
 #import "VSLCallManager.h"
+#import "VSLAudioCodecs.h"
 #import "VSLLogging.h"
 #import "VSLNetworkMonitor.h"
 #import "VSLIpChangeConfiguration.h"
 #import "VSLTransportConfiguration.h"
+#import "VSLVideoCodecs.h"
 
 static NSString * const VSLEndpointErrorDomain = @"VialerSIPLib.VSLEndpoint.error";
 
@@ -82,6 +84,7 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
                 }
                 break;
             }
+            case VSLEndpointClosing:
             case VSLEndpointStarting: {
                 break;
             }
@@ -102,7 +105,7 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
     if (!_networkMonitor) {
         VSLAccount *activeAccount;
         for (VSLAccount *account in self.accounts) {
-            if ([account firstActiveCall]) {
+            if ([self.callManager firstCallForAccount:account]) {
                 activeAccount = account;
                 break;
             }
@@ -134,8 +137,15 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
     VSLLogDebug(@"Creating new PJSIP Endpoint instance.");
     self.state = VSLEndpointStarting;
 
-    // Create PJSUA on the main thread to make all subsequent calls from the main
-    // thread.
+
+    // Create worker thread for pjsip.
+    NSError *threadError;
+    if (![self createPJSIPThreadWithError:&threadError]) {
+        *error = threadError;
+        return NO;
+    }
+
+    // Create PJSUA on the main thread to make all subsequent calls from the main thread.
     pj_status_t status = pjsua_create();
     if (status != PJ_SUCCESS) {
         self.state = VSLEndpointStopped;
@@ -149,12 +159,7 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
         return NO;
     }
 
-    // Create Thread & Pool
-    NSError *threadError;
-    if (![self createPJThreadAndPoolwithError:&threadError]) {
-        *error = threadError;
-        return NO;
-    }
+    [self createPoolForPJSIP];
 
     // Configure the different logging information for the endpoint.
     pjsua_logging_config logConfig;
@@ -198,6 +203,7 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
     mediaConfig.snd_clock_rate = (unsigned int)endpointConfiguration.sndClockRate;
     mediaConfig.has_ioqueue = PJ_TRUE;
     mediaConfig.thread_cnt = 1;
+    mediaConfig.no_vad = PJ_TRUE;
 
     // Initialize Endpoint.
     status = pjsua_init(&endpointConfig, &logConfig, &mediaConfig);
@@ -255,17 +261,22 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
     self.endpointConfiguration = endpointConfiguration;
     self.state = VSLEndpointStarted;
 
-    [self updateCodecs];
+    if (self.endpointConfiguration.codecConfiguration != NULL) {
+        [self updateAudioCodecs];
+        [self updateVideoCodecs];
+    } else {
+        [self updateCodecs];
+    }
 
     return YES;
 }
 
-- (BOOL)createPJThreadAndPoolwithError:(NSError * _Nullable __autoreleasing *)error {
+- (BOOL)createPJSIPThreadWithError:(NSError * _Nullable __autoreleasing *)error {
     // Create a seperate thread
     pj_thread_desc aPJThreadDesc;
     if (!pj_thread_is_registered()) {
         pj_thread_t *pjThread;
-        pj_status_t status = pj_thread_register(NULL, aPJThreadDesc, &pjThread);
+        pj_status_t status = pj_thread_register("VialerPJSIP", aPJThreadDesc, &pjThread);
         if (status != PJ_SUCCESS) {
             if (error != NULL) {
                 *error = [NSError VSLUnderlyingError:nil
@@ -278,24 +289,28 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
         }
     }
 
+    return YES;
+}
+
+- (void)createPoolForPJSIP {
     // Create pool for PJSUA.
     self.pjPool = pjsua_pool_create("VialerSIPLib-pjsua", 1000, 1000);
-    return YES;
 }
 
 - (void)destroyPJSUAInstance {
     VSLLogDebug(@"PJSUA was already running destroying old instance.");
+    self.state = VSLEndpointClosing;
     [self stopNetworkMonitoring];
+    [self.callManager endAllCalls];
 
     for (VSLAccount *account in self.accounts) {
-        [account removeAllCalls];
         [self removeAccount:account];
     }
 
     if (!pj_thread_is_registered()) {
         pj_thread_desc aPJThreadDesc;
         pj_thread_t *pjThread;
-        pj_status_t status = pj_thread_register(NULL, aPJThreadDesc, &pjThread);
+        pj_status_t status = pj_thread_register("VialerPJSIP", aPJThreadDesc, &pjThread);
 
         if (status != PJ_SUCCESS) {
             VSLLogError(@"Error registering thread at PJSUA");
@@ -303,7 +318,7 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
     }
 
     if (self.pjPool != NULL) {
-        pj_pool_release(self->_pjPool);
+        pj_pool_release(self.pjPool);
     }
     
     // Destroy PJSUA.
@@ -325,7 +340,6 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
     NSMutableArray *mutableArray = [self.accounts mutableCopy];
     [mutableArray removeObject:account];
     self.accounts = [mutableArray copy];
-    
 }
 
 - (VSLAccount *)lookupAccount:(NSInteger)accountId {
@@ -345,13 +359,7 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
 }
 
 - (VSLCall *)lookupCall:(NSInteger)callId {
-    for (VSLAccount *account in self.accounts) {
-        VSLCall *call = [account lookupCall:callId];
-        if (call) {
-            return call;
-        }
-    }
-    return nil;
+    return [self.callManager callWithCallId:callId];
 }
 
 - (VSLAccount *)getAccountWithSipAccount:(NSString *)sipAccount {
@@ -446,6 +454,96 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
     return (pj_uint8_t)[priorities[identifier] unsignedIntegerValue];
 }
 
+-(BOOL)updateAudioCodecs {
+    if (self.state != VSLEndpointStarted) {
+        return NO;
+    }
+
+    const unsigned audioCodecInfoSize = 64;
+    pjsua_codec_info audioCodecInfo[audioCodecInfoSize];
+    unsigned audioCodecCount = audioCodecInfoSize;
+    pj_status_t status = pjsua_enum_codecs(audioCodecInfo, &audioCodecCount);
+    if (status != PJ_SUCCESS) {
+        VSLLogError(@"Error getting list of audio codecs");
+        return NO;
+    }
+
+    for (NSUInteger i = 0; i < audioCodecCount; i++) {
+        NSString *codecIdentifier = [NSString stringWithPJString:audioCodecInfo[i].codec_id];
+        pj_uint8_t priority = [self priorityForAudioCodec:codecIdentifier];
+        status = pjsua_codec_set_priority(&audioCodecInfo[i].codec_id, priority);
+        if (status != PJ_SUCCESS) {
+            VSLLogError(@"Error setting codec priority to the correct value");
+            return NO;
+        }
+    }
+    return YES;
+}
+
+-(pj_uint8_t)priorityForAudioCodec:(NSString *)identifier {
+    NSUInteger priority = 0;
+    for (VSLAudioCodecs* audioCodec in self.endpointConfiguration.codecConfiguration.audioCodecs) {
+        if ([VSLAudioCodecString(audioCodec.codec) isEqualToString:identifier]) {
+            priority = audioCodec.priority;
+            return (pj_uint8_t)priority;
+        }
+    }
+    return (pj_uint8_t)priority;
+}
+
+-(BOOL)updateVideoCodecs {
+    if (self.state != VSLEndpointStarted) {
+        return NO;
+    }
+
+    const unsigned videoCodecInfoSize = 64;
+    pjsua_codec_info videoCodecInfo[videoCodecInfoSize];
+    unsigned videoCodecCount = videoCodecInfoSize;
+    pj_status_t videoStatus = pjsua_vid_enum_codecs(videoCodecInfo, &videoCodecCount);
+    if (videoStatus != PJ_SUCCESS) {
+        VSLLogError(@"Error getting list of video codecs");
+        return NO;
+    } else {
+        for (NSUInteger i = 0; i < videoCodecCount; i++) {
+            NSString *codecIdentifier = [NSString stringWithPJString:videoCodecInfo[i].codec_id];
+            pj_uint8_t priority = [self priorityForVideoCodec:codecIdentifier];
+            
+            videoStatus = pjsua_vid_codec_set_priority(&videoCodecInfo[i].codec_id, priority);
+
+            if (priority > 0) {
+                pjmedia_vid_codec_param param;
+                pjsua_vid_codec_get_param(&videoCodecInfo[i].codec_id, &param);
+                param.ignore_fmtp = PJ_TRUE;
+                param.enc_fmt.det.vid.size.w = 288;
+                param.enc_fmt.det.vid.size.h = 352;
+                param.enc_fmt.det.vid.fps.num = 20;
+                param.enc_fmt.det.vid.fps.denum = 1;
+                param.dec_fmt.det.vid.size.w = 1920;
+                param.dec_fmt.det.vid.size.h = 1920;
+                pjsua_vid_codec_set_param(&videoCodecInfo[i].codec_id, &param);
+
+                if (videoStatus != PJ_SUCCESS) {
+                    DDLogError(@"Error setting video codec priority to the correct value");
+                    return NO;
+                }
+            }
+        }
+    }
+
+    return YES;
+}
+
+-(pj_uint8_t)priorityForVideoCodec:(NSString *)identifier {
+    NSUInteger priority = 0;
+    for (VSLVideoCodecs* videoCodec in self.endpointConfiguration.codecConfiguration.videoCodecs) {
+        if ([VSLVideoCodecString(videoCodec.codec) isEqualToString:identifier] && !self.endpointConfiguration.disableVideoSupport) {
+            priority = videoCodec.priority;
+            return (pj_uint8_t)priority;
+        }
+    }
+    return (pj_uint8_t)priority;
+}
+
 #pragma mark - PJSUA callbacks
 
 static void logCallBack(int logLevel, const char *data, int len) {
@@ -458,7 +556,7 @@ static void logCallBack(int logLevel, const char *data, int len) {
     // The data obtained from the callback has a NewLine character at the end, remove it.
     unichar last = [logString characterAtIndex:[logString length] - 1];
     if ([[NSCharacterSet newlineCharacterSet] characterIsMember:last]) {
-        logString = [logString substringToIndex:[logString length]-1];
+        logString = [logString substringToIndex:[logString length] - 1];
     }
 
     switch (logLevel) {
@@ -487,7 +585,7 @@ static void onCallState(pjsua_call_id callId, pjsip_event *event) {
 
     VSLAccount *account = [[VSLEndpoint sharedEndpoint] lookupAccount:callInfo.acc_id];
     if (account) {
-        VSLCall *call = [account lookupCall:callId];
+        VSLCall *call = [[VSLEndpoint sharedEndpoint].callManager callWithCallId:callId];
         if (call) {
             [call callStateChanged:callInfo];
         } else {
@@ -502,7 +600,7 @@ static void onCallMediaState(pjsua_call_id call_id) {
 
     VSLAccount *account = [[VSLEndpoint sharedEndpoint] lookupAccount:callInfo.acc_id];
     if (account) {
-        VSLCall *call = [account lookupCall:call_id];
+        VSLCall *call = [[VSLEndpoint sharedEndpoint].callManager callWithCallId:call_id];
         VSLLogVerbose(@"Received MediaState update for call:%@", call.uuid.UUIDString);
         if (call) {
             [call mediaStateChanged:callInfo];
@@ -525,6 +623,7 @@ static void onRegState2(pjsua_acc_id acc_id, pjsua_reg_info *info) {
     if ([VSLEndpoint sharedEndpoint].ipChangeInProgress) {
         // When disableVideoSupport is on reinivite the calls again. And in the reinivite
         // disable the video stream. Otherwise the response is an 488 Not Acceptatble here.
+        // When videosupport has been disabled in the PBX.
         if ([VSLEndpoint sharedEndpoint].endpointConfiguration.disableVideoSupport) {
             VSLIpChangeConfiguration *ipChangeConfiguration = [VSLEndpoint sharedEndpoint].endpointConfiguration.ipChangeConfiguration;
             VSLAccount *account = [[VSLEndpoint sharedEndpoint] lookupAccount:acc_id];
@@ -585,7 +684,7 @@ static void onTxStateChange(pjsua_call_id call_id, pjsip_transaction *tx, pjsip_
     }
 }
 
-void onIncomingCall(pjsua_acc_id acc_id, pjsua_call_id call_id, pjsip_rx_data *rdata) {
+static void onIncomingCall(pjsua_acc_id acc_id, pjsua_call_id call_id, pjsip_rx_data *rdata) {
     VSLEndpoint *endpoint = [VSLEndpoint sharedEndpoint];
     VSLAccount *account = [endpoint lookupAccount:acc_id];
     if (account) {
@@ -597,20 +696,21 @@ void onIncomingCall(pjsua_acc_id acc_id, pjsua_call_id call_id, pjsip_rx_data *r
                 [VSLEndpoint sharedEndpoint].incomingCallBlock(call);
             }
         }
+        call = nil;
     } else {
         VSLLogWarning(@"Could not accept incoming call. No account found with ID:%d", acc_id);
     }
 }
 
 static void onCallTransferStatus(pjsua_call_id callId, int statusCode, const pj_str_t *statusText, pj_bool_t final, pj_bool_t *continueNotifications) {
-    VSLCall *call = [[VSLEndpoint sharedEndpoint] lookupCall:callId];
+    VSLCall *call = [[VSLEndpoint sharedEndpoint].callManager callWithCallId:callId];
     if (call) {
         [call callTransferStatusChangedWithStatusCode:statusCode statusText:[NSString stringWithPJString:*statusText] final:final == 1];
     }
 }
 
 - (void)callDealloc:(NSNotification *)notification {
-    if (!self.endpointConfiguration.unregisterAfterCall) {
+    if (!self.endpointConfiguration.unregisterAfterCall || self.state != VSLEndpointStarted) {
         return;
     }
 
@@ -655,9 +755,9 @@ static void onCallTransferStatus(pjsua_call_id callId, int statusCode, const pj_
  *  and reinitialize the TCP transport.
  */
 - (void)checkNetworkMonitoring:(NSNotification *)notification {
-    VSLCall *call = notification.userInfo[VSLNotificationUserInfoCallKey];
+    VSLCallState callState = [notification.userInfo[VSLNotificationUserInfoCallStateKey] intValue];
 
-    switch (call.callState) {
+    switch (callState) {
         case VSLCallStateDisconnected: {
             [self stopNetworkMonitoring];
             break;
@@ -665,7 +765,7 @@ static void onCallTransferStatus(pjsua_call_id callId, int statusCode, const pj_
         case VSLCallStateConfirmed: {
             if (!self.monitoringCalls) {
                 for (VSLAccount *account in self.accounts) {
-                    if ([account firstActiveCall]) {
+                    if ([self.callManager firstCallForAccount:account]) {
                         VSLLogVerbose(@"Starting network monitor");
                         [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(ipAddressChanged:) name:VSLNetworkMonitorChangedNotification object:nil];
                         [self.networkMonitor startMonitoring];
@@ -694,7 +794,7 @@ static void onCallTransferStatus(pjsua_call_id callId, int statusCode, const pj_
     BOOL activeCallForAnyAccount = NO;
     // Try to find an account which has an active call.
     for (VSLAccount *account in self.accounts) {
-        if ([account firstActiveCall]) {
+        if ([self.callManager firstCallForAccount:account]) {
             activeCallForAnyAccount = YES;
             break;
         }
@@ -711,7 +811,6 @@ static void onCallTransferStatus(pjsua_call_id callId, int statusCode, const pj_
         [self.networkMonitor stopMonitoring];
         self.networkMonitor = nil;
         self.monitoringCalls = NO;
-        [self callDealloc: nil];
     }
 }
 
@@ -735,7 +834,7 @@ static void onCallTransferStatus(pjsua_call_id callId, int statusCode, const pj_
 static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const pjsua_ip_change_op_info *info) {
     VSLLogInfo(@"onIpChangeProgress:");
 
-    [VSLEndpoint sharedEndpoint].ipChangeInProgress = TRUE;
+    [VSLEndpoint sharedEndpoint].ipChangeInProgress = YES;
 
     switch (op) {
         case PJSUA_IP_CHANGE_OP_RESTART_LIS: {
@@ -756,7 +855,7 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
         }
         case PJSUA_IP_CHANGE_OP_ACC_REINVITE_CALLS: {
             VSLLogInfo(@"Account reinvite calls: %u", status);
-            [VSLEndpoint sharedEndpoint].ipChangeInProgress = FALSE;
+            [VSLEndpoint sharedEndpoint].ipChangeInProgress = NO;
             break;
         }
         default:
@@ -782,8 +881,7 @@ static void onIpChangeProgress(pjsua_ip_change_op op, pj_status_t status, const 
         }
         
         if ([VSLEndpoint sharedEndpoint].missedCallBlock && reason != VSLCallTerminateReasonUnknown) {
-            VSLAccount *account = [[VSLEndpoint sharedEndpoint] lookupAccount:callInfo.acc_id];
-            VSLCall *call = [account lookupCall:call_id];
+            VSLCall *call = [[VSLEndpoint sharedEndpoint].callManager callWithCallId:call_id];;
             if (call) {
                 call.terminateReason = reason;
                 VSLLogDebug(@"Call was terminated for reason: %@", VSLCallTerminateReasonString(reason));